Moving date_limits_to_intervals function out of the main module
